# Docs/API/Api.yaml
openapi: 3.0.3
info:
  title: RAKI API 
  version: "1.0.0-impl"
servers:
  - url: https://api.raki.example.com/v1
...
security:
  - bearerAuth: []
paths:
...
  /merchants/{merchantId}/products/{productId}:
    parameters: [ { $ref: '#/components/parameters/MerchantId' }, { $ref: '#/components/parameters/ProductId' } ]
    get:
      tags: [Products]
      summary: Get product
      responses:
        '200': { description: Product, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [Products]
      summary: Update product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductUpdate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Products]
      summary: Delete product
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
...
  /merchants/{merchantId}/orders/{orderId}:
    parameters: [ { $ref: '#/components/parameters/MerchantId' }, { $ref: '#/components/parameters/OrderId' } ]
    get:
      tags: [Orders]
      summary: Get order with computed totals
      responses:
        '200': { description: Order, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
    patch:
      tags: [Orders]
      summary: Update order (open only) — add/remove items, set discount, requests
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrderUpdate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Orders]
      summary: Cancel order (open only)
      responses:
        '204': { description: Cancelled }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
...
  /merchants/{merchantId}/orders/{orderId}/payments:
    parameters:
      - { $ref: '#/components/parameters/MerchantId' }
      - { $ref: '#/components/parameters/OrderId' }

    post:
      tags: [Payments]
      summary: Record payment (supports split checks via payment_splits)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        '201':
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/ValidationError' }

    get:
      tags: [Payments]
      summary: Get all payments for a specific order
      responses:
        '200':
          description: List of payments for the order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /merchants/{merchantId}/discounts/{discountId}:
    parameters:
      - $ref: '#/components/parameters/MerchantId'
      - $ref: '#/components/parameters/DiscountId'
    get:
      tags: [Discounts]
      summary: Get discount
      responses:
        '200': { description: Discount, content: { application/json: { schema: { $ref: '#/components/schemas/Discount' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }

    patch:
      tags: [Discounts]
      summary: Update discount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DiscountUpdate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Discount' } } } }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }     # locked/used discounts, invalid transitions
        '422': { $ref: '#/components/responses/ValidationError' }

    delete:
      tags: [Discounts]
      summary: Delete discount
      description: Prefer status=inactive when referenced by historical/closed orders; return 409 if hard delete is unsafe.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
...
  /merchants/{merchantId}/reservations/{reservationId}:
    parameters: [ { $ref: '#/components/parameters/MerchantId' }, { $ref: '#/components/parameters/ReservationId' } ]
    patch:
      tags: [Reservations]
      summary: Update or cancel reservation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReservationUpdate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Reservation' } } } }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Reservations]
      summary: Delete individual reservation
      responses:
        '204': { description: Reservation deleted }
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MerchantId: { name: merchantId, in: path, required: true, schema: { type: integer, format: int64 } }
...
    ProductId: { name: productId, in: path, required: true, schema: { type: integer, format: int64 } }
...
    OrderId: { name: orderId, in: path, required: true, schema: { type: integer, format: int64 } }
...
    ReservationId: { name: reservationId, in: path, required: true, schema: { type: integer, format: int64 } }
...
    DiscountId: { name: discountId, in: path, required: true, schema: { type: integer, format: int64 } }
  responses:
    UnauthorizedError:
      description: Unauthenticated (missing/invalid/expired token)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            unauthenticated:
              value: { code: unauthenticated, message: "Missing or invalid access token." }
    ForbiddenError:
      description: Forbidden (valid token, but not allowed by role/tenant)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            forbidden:
              value: { code: forbidden, message: "You do not have permission to perform this action." }
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            not_found:
              value: { code: not_found, message: "Resource not found." }
    BadRequestError:
      description: Bad request (malformed JSON or unsupported value)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            bad_request:
              value: { code: bad_request, message: "Malformed request body." }
    ValidationError:
      description: Validation error (field-level errors)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            validation_error:
              value:
                code: validation_error
                message: "One or more fields are invalid."
                fieldErrors:
                  - field: email
                    message: Email is already in use.
                    code: unique_violation
    ConflictError:
      description: Conflict (concurrency or business rule)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            conflict:
              value: { code: conflict, message: "Conflict with current resource state." }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable code
          enum: [unauthenticated, forbidden, not_found, bad_request, validation_error, conflict, rate_limited, internal]
        message:
          type: string
          description: Human-readable summary
        requestId:
          type: string
          description: Trace id for debugging
        details:
          type: object
          additionalProperties: true
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
              code: { type: string }

    # Merchant/Branch
...
    ProductCreate:
      type: object
      required: [name, tax_category, product_type, merchant_id, base_price, type]
      properties:
        name: { type: string }
        description: { type: string }
        tax_category: { type: integer, format: int64 }
        product_type: { type: integer, format: int64 }
        merchant_id: { type: integer, format: int64 }
        base_price: { type: number, multipleOf: 0.01 }
        type: { type: string, enum: [product, service] }
        status: { type: string, enum: [active, inactive, out_of_stock], default: active }
    ProductUpdate:
      allOf: [ { $ref: '#/components/schemas/ProductCreate' } ]
    Product:
      allOf:
        - $ref: '#/components/schemas/ProductCreate'
        - type: object
          properties:
            product_id: { type: integer, format: int64 }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...
    OrderUpdate:
      type: object
      properties:
        discount_id: { type: integer, format: int64, nullable: true }
        special_requests: { type: string }
        status: { type: string, enum: [open, closed, cancelled, refunded] }
    Order:
      type: object
      properties:
        order_id: { type: integer, format: int64 }
        merchant_id: { type: integer, format: int64 }
        employee_id: { type: integer, format: int64 }
        order_number: { type: string }
        order_date: { type: string, format: date-time }
        order_status: { type: string, enum: [open, closed, cancelled, refunded] }
        discount_id: { type: integer, format: int64, nullable: true }
        special_requests: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        totals:
          type: object
          description: Computed on-the-fly from products + ingredients + discounts + taxes
          properties:
            subtotal: { type: number, multipleOf: 0.01 }
            discount_total: { type: number, multipleOf: 0.01 }
            tax_total: { type: number, multipleOf: 0.01 }
            grand_total: { type: number, multipleOf: 0.01 }
            paid_total: { type: number, multipleOf: 0.01 }
            amount_due: { type: number, multipleOf: 0.01 }
...
    OrderItem:
      type: object
      properties:
        order_item_id: { type: integer, format: int64 }
        order_id: { type: integer, format: int64 }
        product_id: { type: integer, format: int64 }
        quantity: { type: integer }
        payment_status: { type: string, enum: [Unpaid, Paid] }

    # Payments / Splits / Refunds
    PaymentCreate:
      type: object
      required: [payment_method, amount]
      properties:
        payment_method: { type: string, enum: [cash, card, gift_card] }
        amount: { type: number, multipleOf: 0.01 }
        tip_amount: { type: number, multipleOf: 0.01, default: 0.00 }
        split: { type: boolean, default: false }
        stripe_payment_intent_id: { type: string, description: Required for card when integrating Stripe }
    Payment:
      type: object
      properties:
        payment_id: { type: integer, format: int64 }
        order_id: { type: integer, format: int64, nullable: true }
        merchant_id: { type: integer, format: int64 }
        payment_method: { type: string, enum: [cash, card, gift_card] }
        split: { type: boolean }
        status: { type: string, enum: [pending, processing, completed, failed, cancelled, refunded] }
        tip_amount: { type: number, multipleOf: 0.01 }
        amount: { type: number, multipleOf: 0.01 }
        created_at: { type: string, format: date-time }
        processed_at: { type: string, format: date-time, nullable: true }
...
    DiscountCreate:
      type: object
      required: [name, type, value_type, value, start_time]
      properties:
        name: { type: string }
        type: { type: string, enum: [order_discount, product_discount] }
        value_type: { type: string, enum: [value, percentage] }
        value: { type: number }
        minimum_order_value: { type: number, multipleOf: 0.01, default: 0.00 }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [active, inactive, expired], default: inactive }
        max_uses: { type: integer, nullable: true }
    DiscountUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string, enum: [order_discount, product_discount] }
        value_type: { type: string, enum: [value, percentage] }
        value: { type: number }
        minimum_order_value: { type: number, multipleOf: 0.01 }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time, nullable: true }
        status: { type: string, enum: [active, inactive, expired] }
        max_uses: { type: integer, nullable: true }
    Discount:
      allOf:
        - $ref: '#/components/schemas/DiscountCreate'
        - type: object
          properties:
            discount_id: { type: integer, format: int64 }
            current_uses: { type: integer }
            merchant_id: { type: integer, format: int64 }
...
    ReservationCreate:
      type: object
      required: [customer_id, reservation_date, start_time, end_time]
      properties:
        customer_id: { type: integer, format: int64 }
        service_id: { type: integer, format: int64, nullable: true }
        assigned_user_id: { type: integer, format: int64, nullable: true }
        reservation_date: { type: string, format: date }
        start_time: { type: string, pattern: '^\d{2}:\d{2}(:\d{2})?$' }
        end_time: { type: string, pattern: '^\d{2}:\d{2}(:\d{2})?$' }
        deposit_amount: { type: number, multipleOf: 0.01, default: 0.00 }
        total_amount: { type: number, multipleOf: 0.01, default: 0.00 }
    ReservationUpdate:
      type: object
      properties:
        service_id: { type: integer, format: int64, nullable: true }
        assigned_user_id: { type: integer, format: int64, nullable: true }
        reservation_date: { type: string, format: date }
        start_time: { type: string }
        end_time: { type: string }
        status: { type: string, enum: [pending, confirmed, in_progress, completed, cancelled, no_show] }
    Reservation:
      allOf:
        - $ref: '#/components/schemas/ReservationCreate'
        - type: object
          properties:
            reservation_id: { type: integer, format: int64 }
            merchant_id: { type: integer, format: int64 }
            status: { type: string, enum: [pending, confirmed, in_progress, completed, cancelled, no_show] }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...